+++ src/main/java/com/bjfu/fortree/request/woodland/EditWoodlandRequest.java
@@ -63,12 +63,12 @@ public class EditWoodlandRequest {
      * 长度
      */
     @NotNull(message = "样地长度不能为空")
-    private Integer length;
+    private Double length;
     /**
      * 宽度
      */
     @NotNull(message = "样地宽度不能为空")
-    private Integer width;
+    private Double width;
     /**
      * 附加信息(JSON)
      */
